Keep BC w/ old Formatter class since it is likely used by 3rd party plugins. Move Numeric formatter to CoreVisualizations and do not mark w/ @api.
