+++ core/Metrics/Formatter/Html.php
@@ -7,11 +7,13 @@
  */
 namespace Piwik\Metrics\Formatter;
 
+use Piwik\Metrics\Formatter;
+
 /**
  * Metrics formatter that formats for HTML output. Uses non-breaking spaces in formatted values
  * so text will stay unbroken in HTML views.
  */
-class Html extends Api
+class Html extends Formatter
 {
     public function getPrettyTimeFromSeconds($numberOfSeconds, $displayTimeAsSentence = true, $round = false)
     {
